+++ src/os/msg_queue.cc
@@ -89,7 +89,7 @@ ssize_t swMsgQueue_push(swMsgQueue *q, swQueue_data *in, size_t length) {
     return 0;
 }
 
-int swMsgQueue_stat(swMsgQueue *q, int *queue_num, int *queue_bytes) {
+int swMsgQueue_stat(swMsgQueue *q, size_t *queue_num, size_t *queue_bytes) {
     struct msqid_ds __stat;
     if (msgctl(q->msg_id, IPC_STAT, &__stat) == 0) {
         *queue_num = __stat.msg_qnum;
@@ -100,7 +100,7 @@ int swMsgQueue_stat(swMsgQueue *q, int *queue_num, int *queue_bytes) {
     }
 }
 
-int swMsgQueue_set_capacity(swMsgQueue *q, int queue_bytes) {
+int swMsgQueue_set_capacity(swMsgQueue *q, size_t queue_bytes) {
     struct msqid_ds __stat;
     if (msgctl(q->msg_id, IPC_STAT, &__stat) != 0) {
         return -1;
